change strucure like sklearn
