--- pyserini/encode/_collection_iterator.py
@@ -1,71 +1,0 @@
-#
-# Pyserini: Reproducible IR research with sparse and dense representations
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-import json
-import os
-
-from tqdm import tqdm
-
-
-class JsonlCollectionIterator:
-    def __init__(self, collection_path: str, fields=None):
-        if fields:
-            self.fields = fields
-        else:
-            self.fields = ['text']
-        self.all_info = self._load(collection_path)
-        self.size = len(self.all_info['id'])
-        self.batch_size = 1
-        self.shard_id = 0
-        self.shard_num = 1
-
-    def __call__(self, batch_size=1, shard_id=0, shard_num=1):
-        self.batch_size = batch_size
-        self.shard_id = shard_id
-        self.shard_num = shard_num
-        return self
-
-    def __iter__(self):
-        total_len = self.size
-        shard_size = int(total_len / self.shard_num)
-        start_idx = self.shard_id * shard_size
-        end_idx = min(start_idx + shard_size, total_len)
-        if self.shard_id == self.shard_num - 1:
-            end_idx = total_len
-        to_yield = {}
-        for idx in tqdm(range(start_idx, end_idx, self.batch_size)):
-            for key in self.all_info:
-                to_yield[key] = self.all_info[key][idx: min(idx + self.batch_size, end_idx)]
-            yield to_yield
-
-    def _load(self, collection_path):
-        filenames = []
-        if os.path.isfile(collection_path):
-            filenames.append(collection_path)
-        else:
-            for filename in os.listdir(collection_path):
-                filenames.append(os.path.join(collection_path, filename))
-        all_info = {field: [] for field in self.fields}
-        all_info['id'] = []
-        for filename in filenames:
-            with open(filename) as f:
-                for line in tqdm(f):
-                    info = json.loads(line)
-                    all_info['id'].append(str(info['id']))
-                    fields_info = info['contents'].rstrip().split('\n')
-                    for i in range(len(fields_info)):
-                        all_info[self.fields[i]].append(fields_info[i])
-        return all_info
